requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==6.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coverage==6.4.4
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.1
djangorestframework==3.13.1
Faker==12.0.1
filelock==3.8.0
iniconfig==1.1.1
kombu==5.2.4
mixer==7.2.2
packaging==21.3
pika==1.3.0
platformdirs==2.5.2
pluggy==0.3.1
prompt-toolkit==3.0.31
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==3.11
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tox==2.2.1
tzdata==2022.2
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
wrapt==1.14.1
